Revert "Add user account authorized function"
