+++ package.json
@@ -31,8 +31,7 @@
     "animate.css": "^4.1.1",
     "apexcharts": "^3.37.1",
     "axios": "^1.2.1",
-    "crypto-js": "^4.1.1",
-    "daisyui": "^3.2.1",
+    "daisyui": "^2.50.2",
     "dayjs": "^1.11.6",
     "ethers": "^6.1.0",
     "js-base64": "^3.7.5",
