+++ src/components/Modal/account.vue
@@ -1,34 +1,22 @@
 <script setup lang="ts">
+import { useAccountStore } from "@/store/modules/account";
 import { ref } from "vue";
 import { useLang } from "@/hooks/useLang";
 const { t } = useLang();
-import { getNewNostrPrivateKey } from "@/utils/shared";
-import { getPublicKey } from "nostr-tools";
-import { getUserHubContract } from "@/utils/contract/user-hub";
-import { handleEtherError, encrypt } from "@/utils/shared";
+
+defineOptions({
+  name: "account"
+});
 
 const privateKey = ref("");
-const loading = ref(false);
+const accountStore = useAccountStore();
 
-const handleSubmit = async () => {
-  privateKey.value = getNewNostrPrivateKey();
-  const publicKey = getPublicKey(privateKey.value);
-
-  const contract = await getUserHubContract();
-
-  try {
-    loading.value = true;
-    console.log("contract.registerUser", publicKey, encrypt(privateKey.value));
-    const transaction = await contract.registerUser(
-      publicKey,
-      encrypt(privateKey.value)
-    );
-
-    await transaction.wait();
+const handleSubmit = () => {
+  const err = accountStore.savePrivateKey(privateKey.value);
+  if (err) {
+    window.alert(err);
+  } else {
     window.location.reload();
-  } catch (e) {
-    loading.value = false;
-    handleEtherError(e);
   }
 };
 </script>
@@ -43,17 +31,19 @@ const handleSubmit = async () => {
         >✕</label
       >
 
-      <h3 class="font-bold text-lg">{{ t("user.modal title") }}</h3>
-      <p class="py-4">{{ t("user.modal description") }}</p>
+      <h3 class="font-bold text-lg">{{ t("common.input private key") }}</h3>
+      <p class="py-4">
+        <input
+          v-model="privateKey"
+          type="text"
+          :placeholder="t('common.input private key here')"
+          class="input input-bordered input-primary w-full"
+        />
+      </p>
       <div class="modal-action">
-        <label
-          class="btn btn-primary"
-          :class="{ 'btn-disabled': loading }"
-          @click="handleSubmit"
+        <label for="key-modal" class="btn btn-primary" @click="handleSubmit"
+          >{{ t("common.yay") }}!</label
         >
-          <span v-if="loading" class="loading loading-spinner" />
-          <span>{{ t("user.join now") }}!</span>
-        </label>
       </div>
     </div>
   </div>
