+++ src/store/modules/account.ts
@@ -2,17 +2,14 @@ import { defineStore } from "pinia";
 import { store } from "@/store";
 import { accountType } from "./types";
 import { getPublicKey } from "nostr-tools";
-// import { storageLocal } from "@pureadmin/utils";
-import { getUserHubContract } from "@/utils/contract/user-hub";
-import { handleEtherError, decrypt } from "@/utils/shared";
-import { getWalletAddres } from "@/utils/contract/web3";
+import { storageLocal } from "@pureadmin/utils";
 
 export const useAccountStore = defineStore({
   id: "account-settings",
   state: (): accountType => ({
     name: "",
     publicKey: "",
-    privateKey: "" //storageLocal().getItem("sk") || ""
+    privateKey: storageLocal().getItem("sk") || ""
   }),
   getters: {
     getPublicKey() {
@@ -32,19 +29,9 @@ export const useAccountStore = defineStore({
     SET_PRIVATEKEY(key: string) {
       this.privateKey = key;
     },
-    async init(): Promise<any> {
-      try {
-        const address = await getWalletAddres();
-        const contract = await getUserHubContract();
-        const data = await contract.getUserInfo(address);
-        if (!data[0]) return;
-
-        const { privateKey } = data[1];
-        const truePrivateKey = decrypt(privateKey);
-        this.savePrivateKey(truePrivateKey);
-        console.log("2")
-      } catch (e) {
-        handleEtherError(e);
+    init() {
+      if (this.privateKey) {
+        this.savePrivateKey(this.privateKey);
       }
     },
     savePrivateKey(sk: string): string {
@@ -53,7 +40,7 @@ export const useAccountStore = defineStore({
         this.SET_PRIVATEKEY(sk);
         this.SET_PUBLICKEY(pk);
         this.SET_NAME(pk.substring(0, 2));
-        // storageLocal().setItem("sk", sk);
+        storageLocal().setItem("sk", sk);
         return "";
       } catch (e) {
         return e.message;
